Polished incidents + incident dialog +more...
Incident dialog is now reused by incidents belonging to the same
incident chain, disposed once the last incident concludes.
Incidents referenced by currently visible plot branches are now created
as well, which allows to display their consequences next to their
branches, even for random consequences.
Branches are created when the incident to which they belong is selected,
and added to the incident at the last moment.

Added OVDSkin - an extension of the default Skin, which allows to
reference fonts and texture atlases from inside of the skin, even if
they are located in another folder (as long as they can be found via the
file handle resolver)